Rename "legacy" to "desy3"
