+++ txpipe/ingest/__init__.py
@@ -10,4 +10,4 @@ from .ssi import (
     TXIngestSSIDetectionDESBalrog,
 )
 from .dp1 import TXIngestDataPreview1
-from .legacy import TXIngestDESY3Gold
+from .stage3 import TXIngestDESY3Gold
