+++ txpipe/ingest/stage3/desy3.py
@@ -1,5 +1,5 @@
-from ..data_types import HDFFile, MapsFile, FitsFile
-from .base import TXIngestCatalogH5, TXIngestMapsHsp, TXIngestCatalogFits
+from ...data_types import HDFFile, MapsFile, FitsFile
+from ..base import TXIngestCatalogH5, TXIngestMapsHsp, TXIngestCatalogFits
 from ceci.config import StageParameter
